Revert pathing changes to API Ingress
PR #139 correctly fixed a pathing issue for the front-end ingress, but
also made the same changes to the API server ingress in order to
maintain consistency between the two. However, changing the ingress in
the latter case was unnecessary and broke deployments that used
specialized paths for the API server (e.g. using rewrites). This PR
reverts the changes for the API server ingress and leaves the changes to
the front-end ingress untouched.