fix(task-processor): the extraEnv where added in the wrong place
The environment variables added for the task processor should go inside
the container that runs the task processor rather in the deployment
template spec.

The task processor was already including the api environment so it was
not necessary to add merge both `extraEnv`

Also the `extraEnv` keys and values were added as yaml, instead of the
expected object with the `name` and `value` keys.