Disable InfluxDB in values by default
We are no longer using InfluxDB as the default for all new Flagsmith deployments. Disabling it by default in Helm values removes a possible point of friction, for example for customers that require all container images be mirrored in internal repositories.

I understand this should only affect new deployments, but feel free to ignore or close this PR if it would cause any problems.